requirements/dev.txt
django-nose
coverage
factory_boy
